Passes 'user can see account options' spec
