+++ app/controllers/application_controller.rb
@@ -6,4 +6,14 @@ class ApplicationController < ActionController::Base
   def current_user
     @current_user ||= User.find(session[:user_id]) if session[:user_id]
   end
+
+  # def authenticated?
+  #   session[:user_id] && session[:authenticated]
+  # end
+  #
+  # def authenticate!
+  #   redirect_to new_session_path and return unless authenticated?
+  # end
+  #
+  # helper_method :authenticated?
 end
