+++ db/migrate/20170713230517_create_user.rb
@@ -11,5 +11,3 @@ class CreateUser < ActiveRecord::Migration[5.1]
     end
   end
 end
-
-#making sure this shows up
