+++ db/migrate/20170716194448_add_verification_code_to_users.rb
@@ -1,0 +1,5 @@
+class AddVerificationCodeToUsers < ActiveRecord::Migration[5.1]
+  def change
+    add_column :users, :verification_code, :string
+  end
+end
