+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20170714001515) do
+ActiveRecord::Schema.define(version: 20170716194448) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -31,6 +31,7 @@ ActiveRecord::Schema.define(version: 20170714001515) do
     t.string "password_digest"
     t.integer "status"
     t.integer "role"
+    t.string "verification_code"
   end
 
   add_foreign_key "trips", "users"
