+++ lib/code_generator.rb
@@ -1,0 +1,5 @@
+module CodeGenerator
+  def self.generate
+    rand(100000...999999).to_s
+  end
+end
