+++ lib/confirmation_sender.rb
@@ -1,0 +1,7 @@
+module ConfirmationSender
+  def self.send_confirmation_to(user)
+    verification_code = CodeGenerator.generate
+    user.update(verification_code: verification_code)
+    MessageSender.send_code(user.phone, verification_code)
+  end
+end
